Refactor events to models and getters
