+++ hosting/src/views/Event/Event.html
@@ -1,5 +1,15 @@
+<!-- TODO: banner image if exists -->
+<div v-if="false" class="bg-gray-400 w-full">
+  <div class="relative max-w-7xl m-auto overflow-hidden relative">
+    <div class="aspect-w-5 aspect-h-2 w-full overflow-hidden">
+      <div
+        class="absolute bg-cover bg-top inset-0 w-full h-full object-contain"
+        :style="{ backgroundImage: 'url('+event.src+')' }"
+      ></div>
+    </div>
+  </div>
+</div>
+
 <div class="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 mt-8 sm:mt-6">
-  <h3 class="text-3xl font-medium text-brand-black pb-8">
-    // {{ $route.params.slug }} //
-  </h3>
+  <h3 class="text-3xl font-medium text-brand-black pb-8">{{ event.name }}</h3>
 </div>
